updates per coments
